+++ hpc/nvidia/cuda/fakeCUDA/cuda.cpp
@@ -1,27 +1,8 @@
 // SPDX-License-Identifier: GPL-3.0
 /* Copyright (c) 2025 Rong Tao */
-#ifdef HAVE_HPCC
-#include <hc_runtime.h>
-#include <hcc/hcc_internal.h>
-#include <cuda_adapter.h>
-#else
-#include <cuda.h>
-#include <cuda_runtime.h>
-#endif
 #include "device.h"
+#include "cuda.hpp"
 
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-#ifdef HAVE_HPCC
-# define CUresult	hcError_t
-# undef CUDA_ERROR_INVALID_VALUE
-# define CUDA_ERROR_INVALID_VALUE	cudaErrorInvalidValue
-# undef CUDA_SUCCESS
-# define CUDA_SUCCESS	cudaSuccess
-#endif
 
 CUresult cuDeviceGet(CUdevice *device, int ordinal)
 {
@@ -46,6 +27,3 @@ CUresult cuDeviceComputeCapability(int *major, int *minor, CUdevice dev)
 
 	return CUDA_SUCCESS;
 }
-#ifdef __cplusplus
-}
-#endif