+++ hpc/nvidia/cuda/fakeCUDA/cuda.hpp
@@ -1,0 +1,36 @@
+// SPDX-License-Identifier: GPL-3.0
+/* Copyright (c) 2025 Rong Tao */
+#ifdef HAVE_HPCC
+#include <hc_runtime.h>
+#include <hcc/hcc_internal.h>
+#include <cuda_adapter.h>
+#else
+#include <cuda.h>
+#include <cuda_runtime.h>
+#endif
+#include "device.h"
+
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+#ifdef HAVE_HPCC
+# define CUresult	hcError_t
+# undef CUDA_ERROR_INVALID_VALUE
+# define CUDA_ERROR_INVALID_VALUE	cudaErrorInvalidValue
+# undef CUDA_SUCCESS
+# define CUDA_SUCCESS	cudaSuccess
+#endif
+
+CUresult cuDeviceGet(CUdevice *device, int ordinal);
+
+CUresult cuDeviceComputeCapability(int *major, int *minor, CUdevice dev);
+
+CUresult cuModuleLoad(CUmodule* module, const char *fname);
+
+CUresult cuModuleGetFunction(CUfunction* hfunc, CUmodule hmod, const char *name);
+
+#ifdef __cplusplus
+}
+#endif