+++ hpc/nvidia/cuda/fakeCUDA/hccl.hpp
@@ -1,0 +1,16 @@
+// SPDX-License-Identifier: GPL-3.0
+/* Copyright (c) 2025 Rong Tao */
+#include <hccl.h>
+#include <hc_runtime.h>
+#include <hcc/hcc_internal.h>
+
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+hcclResult_t hcclGetVersion(int *version);
+
+#ifdef __cplusplus
+}
+#endif