+++ hpc/nvidia/cuda/fakeCUDA/sparse.cpp
@@ -1,19 +1,9 @@
 // SPDX-License-Identifier: GPL-3.0
 /* Copyright (c) 2025 Rong Tao */
-#ifdef HAVE_HPCC
-#include <hc_runtime.h>
-#include <hcsparse/hcsparse.h>
-#include <cuda_adapter.h>
-#else
-#define DISABLE_CUSPARSE_DEPRECATED	1
-#include <cusparse.h>
-#endif
+#include "sparse.hpp"
 #include "debug.h"
 
 
-#ifdef __cplusplus
-extern "C" {
-#endif
 cusparseStatus_t cusparseGetProperty(libraryPropertyType type, int *value)
 {
 	DEBUG_DBG("type %d\n", type);
@@ -33,6 +23,3 @@ cusparseStatus_t cusparseGetProperty(libraryPropertyType type, int *value)
 	}
 	return CUSPARSE_STATUS_SUCCESS;
 }
-#ifdef __cplusplus
-}
-#endif